+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Framer Motion adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.4] 2020-07-24
+
+### Added
+
+-   Drag Cypress test suite.
+
+### Fixed
+
+-   Application of relative drag constraints.
+
 ## [2.1.3] 2020-07-23
 
 ### Fixed